+++ example/pages/toast.vue
@@ -3,6 +3,7 @@
     <wv-button type="default" @click="showSuccess">成功提示</wv-button>
     <wv-button type="default" @click="showError">出错提示</wv-button>
     <wv-button type="default" @click="showText">纯文本提示</wv-button>
+    <wv-button type="default" @click="showLoading">加载提示</wv-button>
   </div>
 </template>
 
@@ -12,7 +13,9 @@
   export default {
     methods: {
       showSuccess () {
-        Toast('操作成功')
+        Toast('操作成功').then(() => {
+          console.log('closed')
+        })
       },
 
       showError () {
@@ -20,11 +23,21 @@
           duration: 1000,
           message: '操作失败',
           icon: 'warn'
+        }).then(() => {
+          console.log('closed')
         })
       },
 
       showText () {
         Toast({
+          duration: 1000,
+          message: 'hello',
+          type: 'text'
+        })
+      },
+
+      showLoading () {
+        Toast.loading({
           duration: 1000,
           message: 'hello',
           type: 'text'